Changes: **Introducing Encryption for password using Spring Security**
For that purpose we used BCryptEncoder for encoding, created an instance
of BCryptEncoder in `UsersServiceImpl` to initialize it using
constructor.
As BCryptPasswordEncoder is a regular class and not defined as a
component or a bean.
So for us to be able to inject an object of this class into user service
implementation, we have created a bean in the form of a method in
`PhotoAppUsersApplication` class.