Changes: **Configuring Spring Security to allow HTTP requests from certain IP Address**
Here I will accept the requests from only the IP address of our
API/Cloud gateway.
We will use `access` method
Access method can accept web security expressions and the epression that
w want to use is called `hasIpAddress`.
Way to create it is `new
WebExpressionAuthorizationManager("hasIpAddress('"+environment.getProperty("gateway.ip")+"')")`
, this is the ip address where API gateway application is running.
Put this IP address in properties file and get it from there using
`@Value` annotation
And we will need to use the `Environment` class instance to get the
value from the application.properties file before the bean creation.

The @Value("${property.name}") annotation will inject the value into the
field after the bean is created, but before the bean is fully
initialized.

In the case of the WebSecurity class, the @Value("${gateway.ip}")
annotation will inject the value into the gatewayIp field after the bean
is created, but before the configure(HttpSecurity http) method is
called.

However, there is a catch. The @Value annotation is processed by the
Spring bean post-processor, which is executed after the bean is created,
but before the bean is fully initialized.

In the case of the WebSecurity class, the configure(HttpSecurity http)
method is annotated with @Bean, which means it is executed during the
bean creation process. Specifically, it is executed during the "bean
definition" phase, which is before the "bean post-processing" phase.

So, in this case, the @Value("${gateway.ip}") annotation will not have
injected the value into the gatewayIp field yet when the
configure(HttpSecurity http) method is called.

On the other hand, when you inject the Environment object into the
constructor of the WebSecurity class, you can access the property value
immediately, because the constructor is executed during the bean
creation process, before the bean is fully initialized.

So, to answer your question, the @Value("${property.name}") annotation
will not be able to inject the value before the execution of the
configure(HttpSecurity http) method, because the @Value annotation is
processed after the bean creation process.

In contrast, injecting the Environment object into the constructor of
the WebSecurity class allows you to access the property value
immediately, before the configure(HttpSecurity http) method is called.

Therefore, in this case, injecting the Environment object into the
constructor of the WebSecurity class is a better approach than using the
@Value("${property.name}") annotation.